Add cpt.load_mesh and cpt.load_from_meshio functions.
