+++ docs/changelog.rst
@@ -19,6 +19,13 @@ Major changes
   A missing key element has been added and the :class:`~capytaine.green_functions.delhommeau.XieDelhommeau` is now actually more accurate near the free surface.
   (:pull:`180` and :pull:`216`)
 
+Minor changes
+~~~~~~~~~~~~~
+
+* :func:`~capytaine.io.mesh_loaders.load_mesh` is now exported by the main namespace, that is available with :code:`from capytaine import load_mesh`.
+  A function :func:`~capytaine.io.meshio.load_from_meshio` is also now exported in the main namespace.
+  The documentation has been changed to recommend the use of these functions instead of :meth:`~capytaine.bodies.bodies.FloatingBody.from_file` and :meth:`~capytaine.bodies.bodies.FloatingBody.from_meshio`.
+
 Internals
 ~~~~~~~~~
